requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Limiter
uuid6
gdown
PyJWT
httpx
python-dotenv
pytest